Fix pondering followed by stop immediately
Sometimes the engine is asked to ponder, and then stop almost immediately, before the engine
has a move in mind. Which was causing it to send an empty move (a1a1), which makes some GUIs
crash (LiChess Bot for example). A work around is to pick a move randomly. This might backfire
but at least won't make the engine/GUI crash