Replace the net with a new one
The outgoing net had some traces of impure data, this one fixes taht
I verified that all the data is from a net that was purely trained
on Zahak self-games, and all the games are from Zahak self-games

The strength is about equal to what the outgoing net has. Data size was 213M FENs.